requirements.txt
bcrypt==3.1.5
blinker==1.4
certifi==2018.11.29
cffi==1.11.5
chardet==3.0.4
Click==7.0
DateTime==4.3
Flask==1.0.2
Flask-Mail==0.9.1
gunicorn==19.9.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.1.0
onesignal-sdk==0.1.1
pycparser==2.19
PyJWT==1.7.1
pymongo==3.4.0
pytz==2018.9
requests==2.21.0
six==1.12.0
urllib3==1.24.1
Werkzeug==0.14.1
zope.interface==4.6.0
